Completed Active Record models for posts and authors, added association and validation. Added tests for association and validation.
