+++ Gemfile
@@ -1,5 +1,7 @@
 source 'https://rubygems.org'
 
+#For testing
+gem 'shoulda'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 gem 'rails', '4.2.3'
@@ -42,4 +44,3 @@ group :development, :test do
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
 end
-
