+++ db/migrate/20150728172528_create_authors.rb
@@ -1,0 +1,11 @@
+class CreateAuthors < ActiveRecord::Migration
+  def change
+    create_table :authors do |t|
+      t.string :first_name
+      t.string :last_name
+      t.string :bio
+
+      t.timestamps null: false
+    end
+  end
+end
