+++ db/migrate/20150728182702_rem_author_from_post.rb
@@ -1,0 +1,5 @@
+class RemAuthorFromPost < ActiveRecord::Migration
+  def change
+    remove_column :posts, :author
+  end
+end
