+++ test/fixtures/authors.yml
@@ -1,0 +1,11 @@
+# Read about fixtures at http://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
+
+one:
+  first_name: MyString
+  last_name: MyString
+  bio: MyString
+
+two:
+  first_name: MyString
+  last_name: MyString
+  bio: MyString
