+++ test/fixtures/posts.yml
@@ -1,0 +1,17 @@
+# Read about fixtures at http://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
+
+one:
+  title: MyString
+  body: MyText
+  summary: MyString
+  author: MyString
+  created_at: MyString
+  author_id: 1
+
+two:
+  title: MyString
+  body: MyText
+  summary: MyString
+  author: MyString
+  created_at: MyString
+  author_id: 1
